requirements.txt
langchain
langchain-groq
streamlit
python-dotenv
langchain_community
pypdf
jinja2